todoproject/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.2
